requirements.txt
Jinja2==3.1.5
MarkupSafe==2.1.5
PyYAML==6.0.2
